# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Shovel Heroes Public API
  version: 0.1.0
  description: |
    REST endpoints replacing previous Base44 SDK usage.
    This spec is inferred from the frontend code. Adjust field types and required properties to match your backend.
...
servers:
  - url: https://your.api.server
    description: Production
  - url: http://localhost:8787
    description: Local Dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ID:
      type: string
      description: 唯一識別字串 (UUID、ULID 或資料庫主鍵)
      example: "grid_01HZYQ9W123ABCDEF"
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 時間戳記 (UTC)
      example: "2025-10-02T08:12:30Z"
    DisasterArea:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
          description: 災區名稱
          example: 光復鄉重災區
        center_lat:
          type: number
          format: double
          description: 災區中心點緯度
          example: 23.8751
        center_lng:
          type: number
          format: double
          description: 災區中心點經度
          example: 121.578
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
      required: [id, name, center_lat, center_lng]
    SupplyNeed:
      type: object
      properties:
        name:
          type: string
          description: 需求物資名稱
          example: 鋤頭
        quantity:
          type: integer
          description: 需求總數量
          example: 20
        unit:
          type: string
          description: 數量單位
          example: 支
        received:
          type: integer
          description: 已收到的數量
          default: 0
          example: 5
      required: [name, quantity, unit]
    Grid:
      type: object
      properties:
        id:
            $ref: '#/components/schemas/ID'
        code:
          type: string
          description: 網格代碼 (在災區中應唯一)
          example: A-3
        grid_type:
          type: string
          description: 網格類型 (決定使用情境)
          enum: [mud_disposal, manpower, supply_storage, accommodation, food_area]
          example: manpower
        disaster_area_id:
          $ref: '#/components/schemas/ID'
        volunteer_needed:
          type: integer
          description: 需要志工數量 (僅 manpower 類型有意義)
          default: 0
          example: 10
        volunteer_registered:
          type: integer
          description: 已報名志工數量
          default: 0
          example: 4
        meeting_point:
          type: string
          nullable: true
          description: 集合地點描述
          example: 光復國小正門
        risks_notes:
          type: string
          nullable: true
          description: 風險或安全注意事項
          example: 地面濕滑，注意積水與電線
        contact_info:
          type: string
          nullable: true
          description: 格主或聯絡人資訊 (手機/Line)
          example: 0912-345-678
        center_lat:
          type: number
          format: double
          description: 中心點緯度
          example: 23.87525
        center_lng:
          type: number
          format: double
          description: 中心點經度
          example: 121.57812
        bounds:
          type: object
          description: 粗略網格邊界 (地圖顯示用)
          properties:
            north:
              type: number
              example: 23.876
            south:
              type: number
              example: 23.874
            east:
              type: number
              example: 121.579
            west:
              type: number
              example: 121.577
          required: [north, south, east, west]
        status:
          type: string
          description: 網格狀態
          enum: [open, closed, completed, pending]
          default: open
          example: open
        supplies_needed:
          type: array
          description: 所需物資列表
          items:
            $ref: '#/components/schemas/SupplyNeed'
          example:
            - name: 鋤頭
              quantity: 20
              unit: 支
              received: 5
            - name: 工作用手套
              quantity: 50
              unit: 副
              received: 10
        grid_manager_id:
          $ref: '#/components/schemas/ID'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
      required: [id, code, grid_type, disaster_area_id, center_lat, center_lng, bounds, status]
...
    SupplyDonation:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        grid_id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
          description: 捐贈物資名稱
          example: 礦泉水
        quantity:
          type: integer
          description: 捐贈數量
          example: 100
        unit:
          type: string
          description: 單位
          example: 箱
        donor_contact:
          type: string
          description: 捐贈者聯絡方式
          example: water-donor@example.com
        created_at:
          $ref: '#/components/schemas/Timestamp'
      required: [id, grid_id, name, quantity, unit]
    GridDiscussion:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        grid_id:
          $ref: '#/components/schemas/ID'
        user_id:
          $ref: '#/components/schemas/ID'
        content:
          type: string
          description: 留言內容
          example: 今天下午 2 點再集中一次清運。
        created_at:
          $ref: '#/components/schemas/Timestamp'
      required: [id, grid_id, user_id, content]
    Announcement:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        title:
          type: string
          description: 公告標題
          example: 10/3 志工集合時間調整
        body:
          type: string
          description: 公告內容 (Markdown 可選)
          example: 明日集合時間改為 **08:30**，請提早 10 分鐘報到。
        created_at:
          $ref: '#/components/schemas/Timestamp'
      required: [id, title, body]
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: 錯誤訊息
          example: 未授權的請求
        code:
          type: string
          description: 內部錯誤代碼 (可選)
          example: AUTH_REQUIRED
      required: [message]
  parameters:
    PageLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: 單頁筆數 (預設 50, 上限 200)
    PageOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: 起始位移 (用於分頁)
  responses:
    Unauthorized:
      description: 未授權 (需要 Bearer Token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 找不到資源
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 請求參數或內容驗證失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
  /disaster-areas:
    get:
      summary: List disaster areas
      description: 取得所有災區清單 (含中心座標) 。
      tags: [DisasterAreas]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DisasterArea' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      summary: Create disaster area
      description: 新增一筆災區資料。
      tags: [DisasterAreas]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisasterArea'
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /disaster-areas/{id}:
    get:
      summary: Get disaster area
      description: 依 ID 取得單一災區。
      tags: [DisasterAreas]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/ID' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DisasterArea' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      summary: Update disaster area
      description: 更新災區內容。
      tags: [DisasterAreas]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/ID' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DisasterArea' }
      responses:
        '200': { description: Updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      summary: Delete disaster area
      description: 刪除指定災區 (需權限)。
      tags: [DisasterAreas]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/ID' }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /grids:
    get:
      summary: List grids
      description: 取得全部網格 (可於未來加入篩選參數)。
      tags: [Grids]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Grid' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      summary: Create grid
      description: 建立新的救援 / 資源網格。
      tags: [Grids]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Grid' }
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /grids/{id}:
    get:
      summary: Get grid
      description: 依 ID 取得單一網格詳情。
      tags: [Grids]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/ID' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Grid' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      summary: Update grid
      description: 更新網格資料 (含需求/狀態)。
      tags: [Grids]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/ID' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Grid' }
      responses:
        '200': { description: Updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      summary: Delete grid
      description: 刪除網格 (通常僅管理員使用)。
      tags: [Grids]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/ID' }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /supply-donations:
    get:
      summary: List supply donations
      description: 取得物資捐贈紀錄清單。
      tags: [SupplyDonations]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SupplyDonation' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      summary: Create supply donation
      description: 建立一筆捐贈物資紀錄。
      tags: [SupplyDonations]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplyDonation' }
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /grid-discussions:
    get:
      summary: List grid discussions
      description: 取得全部網格的留言 (建議未來加入過濾)。
      tags: [GridDiscussions]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/GridDiscussion' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      summary: Post discussion message
      description: 在指定網格張貼留言。
      tags: [GridDiscussions]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GridDiscussion' }
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /announcements:
    get:
      summary: List announcements
      description: 取得公告列表 (可依時間排序)。
      tags: [Announcements]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Announcement' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      summary: Create announcement
      description: 建立一則公告。
      tags: [Announcements]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Announcement' }
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /functions/fix-grid-bounds:
    post:
      summary: Normalize / repair stored grid bounds
      description: 批次修正網格邊界 (僅管理維運)。
      tags: [Functions]
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: Done }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /functions/export-grids-csv:
    get:
      summary: Export grids as CSV
      description: 匯出所有網格資料為 CSV 檔。
      tags: [Functions]
      responses:
        '200':
          description: CSV content
          content:
            text/csv:
              schema: { type: string, description: CSV text }
        '500': { $ref: '#/components/responses/InternalError' }
  /functions/import-grids-csv:
    post:
      summary: Import grids from CSV content
      description: 從 CSV 內容匯入 / 更新網格。
      tags: [Functions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                csv: { type: string }
              required: [csv]
      responses:
        '200': { description: Import result }
        '400': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalError' }
  /functions/grid-template:
    get:
      summary: Download blank grid template CSV
      description: 下載空白網格匯入範本。
      tags: [Functions]
      responses:
        '200': { description: CSV }
        '500': { $ref: '#/components/responses/InternalError' }
  /functions/external-grid-api:
    post:
      summary: Proxy / external grid API call
      description: 代理對外部網格系統的呼叫。
      tags: [Functions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: OK }
        '400': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalError' }
  /functions/external-volunteer-api:
    post:
      summary: Proxy / external volunteer API call
      description: 代理對外部志工系統的呼叫。
      tags: [Functions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: OK }
        '400': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalError' }
...